requirements.txt
# Core
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
pymysql>=1.1.0
cryptography>=41.0.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Resume parsing
pdfplumber>=0.10.0
python-docx>=1.1.0
PyMuPDF>=1.23.0

# Text processing & evaluation
rapidfuzz>=3.5.0

# LLM (optional)
zhipuai>=2.0.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.2

